backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-dotenv
requests
psycopg2-binary
# NEW ↓
pytesseract
Pillow
pdf2image